+++ libraries/dashboard_api.rb
@@ -24,8 +24,6 @@ module GrafanaCookbook
       )
 
       JSON.parse(response.body)
-    rescue BackendError
-      nil
     end
 
     #